refactor: extract onboarding step components (#7)
Split 1000-LOC onboarding page into 4 focused sub-components:
- OnboardingStepClinic (clinic creation form)
- OnboardingStepStaff (staff member management)
- OnboardingStepServices (service configuration)
- OnboardingStepTimeSlots (time slot scheduling)

Main page reduced from ~1000 to ~505 LOC.
